modify recipe apis

+++ recipe/views.py
@@ -13,7 +13,13 @@ from drf_yasg import openapi
 @permission_classes([permissions.IsAuthenticated])
 def recipe_list(request):
     if request.method == "GET":
-        recipes = Recipe.objects.all()
+        if request.user.is_staff:
+            # If user is staff, fetch all recipes
+            recipes = Recipe.objects.all()
+        else:
+            # If user is not staff, fetch only their own recipes
+            recipes = Recipe.objects.filter(owner=request.user)
+
         serializer = RecipeSerializer(recipes, many=True)
         return Response(serializer.data)
 
@@ -29,17 +35,24 @@ def recipe_list(request):
 @api_view(["GET", "PUT", "DELETE"])
 @permission_classes([permissions.IsAuthenticated, IsRecipeOwnerOrAdmin])
 def recipe_detail(request, pk):
-    print(pk)
     try:
         recipe = Recipe.objects.get(pk=pk)
     except Recipe.DoesNotExist:
         return Response({"error": "Recipe not found"}, status=status.HTTP_404_NOT_FOUND)
+    except Exception as e:
+        return Response({"error": str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
     if request.method == "GET":
         serializer = RecipeSerializer(recipe)
         return Response(serializer.data)
 
     elif request.method == "PUT":
+        # Check if the user is a staff member or the owner of the recipe
+        if not request.user.is_staff and request.user != recipe.owner:
+            return Response(
+                {"error": "Permission denied"}, status=status.HTTP_403_FORBIDDEN
+            )
+
         serializer = RecipeSerializer(recipe, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -49,5 +62,11 @@ def recipe_detail(request, pk):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == "DELETE":
+        # Check if the user is a staff member or the owner of the recipe
+        if not request.user.is_staff and request.user != recipe.owner:
+            return Response(
+                {"error": "Permission denied"}, status=status.HTTP_403_FORBIDDEN
+            )
+
         recipe.delete()
         return Response({"message": "Successfully deleted"}, status=status.HTTP_200_OK)
